Latest work on both the parser and the gui
Initial work on the parser (Requires additional testing and parsing)

Fixed an issue with the GUI throwing an exception as it was trying to
call an internal method. The reason was that I renamed the project name
later on, but the project configuration still had the original name.
During runtime, the application was trying to load methods from a
compiled assembly that didn't exist. I found the configuration and named
it appropriately, and no longer have the error.