Parsing improved and more data collected
I now have the order in which is collects the data improved. Previously,
I ran into some identifiers which weren't expected and I was collecting
too much.

Improved the output report to reflect the correct column order and data.

Need to test alias checking because it appears my sample doesn't include
it for the host or subnet configurations.

Nat parsing needs to still be completed.