Phase 1 complete
Completed work to parse out needed information for object networks.
Includes Nat parsing and alias resolving. If an alias can't be resolved,
a comment will be provided to explain this. Additionally, I also include
parsing for port information since this was included in a number of the
configuration examples. See source\samples for files you can use in the
project. To do this, run the application in the "app" folder and
copy/paste the data in the sample file(s) into the app. Then, click
"Generate" to start the process. When it's complete, it will write the
new file to disc and open it for you.