Added support to write data to xml file - Version 1.0.0.3
This may or may not improve the experience of seeing the data. The
program will output both csv and xml data.

While the csv file allows a nice "quick and dirty" view of the data, the
xml file will show a much cleaner dynamic list that renders nicely in MS
Excel.