Refactoring and Initial Parsing / Output of Access Lists
Refactored the design for managing both access lists and object networks
to avoid duplicate work and bloat.
Added support for parsing access lists to file. Currently only parsing
the acl types "standard" and "remarks" (Low hanging fruit in the task
list and provides a proof of concept). Extended is next and last on the
list, will be much more involved.

Improved the completion experience for the user now that we are
involving 2 different files. User will receive a confirmation popup
(Much like before) with a cleaner message explaining what files were
created. Once the user closes the message box, the selected directory
will be open in Windows Explorer to view the files.

Incremented the version number to 1.0.0.5