Completed parsing for ACL Extended types
Extended types are complete. Unfortunately, there were a few conditions
which I had to micromanage until I can see a better pattern. However,
seems to be doable at this time. I also completed the ip alias
translation as well.

Incremented the version to 1.1.0.0 with the completion of this new
feature.