Minor GUI changes and added to unit tests
Added menu option to consolidate the buttons.
Removed the event to open the output folder at the end (There is a menu
option to open this if needed).
Files are now opened (Based on request) in the following order:
AccessList then ObjectNetwork. The point was to ensure that the object
network list was the one they see first after all is opened.

Version incremented to 1.1.0.1