New Parsing Conditions and Errors
1. Fixing original issue with had to do with an out of sequence error
parsing a string when the index was -1. The value was not an expected
condition (Value: "!") which has been fixed going forward.

2. Found other conditions which break the business logic discussed for
parsing access-lists. We will need to discuss how they fit the expected
functionality. However, I added better handling and logging to output to
the log file all lines which failed (Should have been there in the first
place) and continue parsing lines which fit the logic. This will help
avoid show stoppers and still provide communication for why certain
lines couldn't be parsed.

Incremented Version